There is no need for an interface and an implementation that are both in the same directory.
